Refactoring to use DMNProfile as FEELProfile(s)
